Extract date of birth form for life page
